requirements.txt
python-chess>=1.11.2
psycopg2-binary>=2.9.9
requests>=2.31.0

